geofence fix 2
