--- app/src/main/java/com/cs4518/poseidon/myapplication/GeofenceBroadcastReceiver.java
@@ -1,71 +1,0 @@
-package com.cs4518.poseidon.myapplication;
-
-import android.content.BroadcastReceiver;
-import android.app.NotificationManager;
-import android.app.PendingIntent;
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.Intent;
-import android.graphics.BitmapFactory;
-import android.media.AudioManager;
-import android.support.v4.app.TaskStackBuilder;
-import android.support.v7.app.NotificationCompat;
-import android.util.Log;
-
-import com.google.android.gms.location.Geofence;
-import com.google.android.gms.location.GeofencingEvent;
-
-/**
- * Created by jiuchuan on 2/8/18.
- */
-
-public class GeofenceBroadcastReceiver extends BroadcastReceiver {
-
-    public static final String TAG = GeofenceBroadcastReceiver.class.getSimpleName();
-
-    public boolean inGeofencing = false;
-
-    @Override
-    public void onReceive(Context context, Intent intent) {
-        // Get the Geofence Event from the Intent sent through
-        GeofencingEvent geofencingEvent = GeofencingEvent.fromIntent(intent);
-        if (geofencingEvent.hasError()) {
-            Log.e(TAG, String.format("Error code : %d", geofencingEvent.getErrorCode()));
-            return;
-        }
-
-        // Get the transition type.
-        int geofenceTransition = geofencingEvent.getGeofenceTransition();
-        // Check which transition type has triggered this event
-        if (geofenceTransition == Geofence.GEOFENCE_TRANSITION_ENTER) {
-            inGeofencing = true;
-        } else if (geofenceTransition == Geofence.GEOFENCE_TRANSITION_EXIT) {
-            inGeofencing = false;
-        } else {
-            // Log the error.
-            Log.e(TAG, String.format("Unknown transition : %d", geofenceTransition));
-            // No need to do anything else
-            return;
-        }
-    }
-
-
-
-    /**
-     * Changes the ringer mode on the device to either silent or back to normal
-     *
-     * @param context The context to access AUDIO_SERVICE
-     * @param mode    The desired mode to switch device to, can be AudioManager.RINGER_MODE_SILENT or
-     *                AudioManager.RINGER_MODE_NORMAL
-     */
-    private void setRingerMode(Context context, int mode) {
-        NotificationManager nm = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
-        // Check for DND permissions for API 24+
-        if (android.os.Build.VERSION.SDK_INT < 24 ||
-                (android.os.Build.VERSION.SDK_INT >= 24 && !nm.isNotificationPolicyAccessGranted())) {
-            AudioManager audioManager = (AudioManager) context.getSystemService(Context.AUDIO_SERVICE);
-            audioManager.setRingerMode(mode);
-        }
-    }
-
-}
