--- app/src/main/java/com/cs4518/poseidon/myapplication/Geofencing.java
@@ -1,127 +1,0 @@
-package com.cs4518.poseidon.myapplication;
-
-import android.app.PendingIntent;
-import android.content.Context;
-import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.util.Log;
-
-import com.google.android.gms.common.api.GoogleApiClient;
-import com.google.android.gms.common.api.Result;
-import com.google.android.gms.common.api.ResultCallback;
-import com.google.android.gms.location.Geofence;
-import com.google.android.gms.location.GeofencingClient;
-import com.google.android.gms.location.GeofencingRequest;
-import com.google.android.gms.location.LocationServices;
-import com.google.android.gms.location.places.Place;
-import com.google.android.gms.location.places.PlaceBuffer;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * @author jiuchuan
- * @version Feb 8, 2018
- */
-
-public class Geofencing implements OnCompleteListener<Void> {
-
-    // Constants
-    public static final String TAG = Geofencing.class.getSimpleName();
-    private static final float GEOFENCE_RADIUS = 50;
-    private static final long GEOFENCE_TIMEOUT = 24 * 60 * 60 * 1000;
-
-    private List<Geofence> mGeofenceList;
-    private PendingIntent mGeofencePendingIntent;
-    private GoogleApiClient mGoogleApiClient;
-    private Context mContext;
-
-    public Geofencing(Context context, GoogleApiClient client){
-        mContext = context;
-        mGoogleApiClient = client;
-        mGeofencePendingIntent = null;
-        mGeofenceList = new ArrayList<>();
-    }
-
-    public void registerAllGeofences(){
-        if (mGoogleApiClient == null || !mGoogleApiClient.isConnected() ||
-                mGeofenceList == null || mGeofenceList.size() == 0){
-            return;
-        }
-        try{
-            GeofencingClient geofencingClient = LocationServices.getGeofencingClient(mContext);
-            geofencingClient
-                    .addGeofences(getGeofencingRequest(), getGeofencePendingIntent())
-                    .addOnCompleteListener(this);
-//            LocationServices.GeofencingApi.addGeofences(
-//                    mGoogleApiClient,
-//                    getGeofencingRequest(),
-//                    getGeofencePendingIntent()
-//            ).setResultCallback(this);
-        } catch (SecurityException securityException) {
-
-        }
-    }
-
-    @Override
-    public void onComplete(@NonNull Task<Void> task) {
-//        mPendingGeofenceTask = PendingGeofenceTask.NONE;
-//        if (task.isSuccessful()) {
-//            updateGeofencesAdded(!getGeofencesAdded());
-//            setButtonsEnabledState();
-//
-//            int messageId = getGeofencesAdded() ? R.string.geofences_added :
-//                    R.string.geofences_removed;
-//            Toast.makeText(this, getString(messageId), Toast.LENGTH_SHORT).show();
-//        } else {
-//            // Get the status code for the error and log it using a user-friendly message.
-//            String errorMessage = GeofenceErrorMessages.getErrorString(this, task.getException());
-//            Log.w(TAG, errorMessage);
-//        }
-    }
-
-    public void updateGeofencesList(PlaceBuffer places){
-        mGeofenceList = new ArrayList<>();
-        if (places == null || places.getCount() == 0) return;
-        for (Place place : places) {
-            String placeUID = place.getId();
-            double placeLat = place.getLatLng().latitude;
-            double placeLng = place.getLatLng().longitude;
-            //Build a Geofence object
-            Geofence geofence = new Geofence.Builder()
-                    .setRequestId(placeUID)
-                    .setExpirationDuration(GEOFENCE_TIMEOUT)
-                    .setCircularRegion(placeLat, placeLng, GEOFENCE_RADIUS)
-                    .setTransitionTypes(Geofence.GEOFENCE_TRANSITION_ENTER | Geofence.GEOFENCE_TRANSITION_EXIT)
-                    .build();
-            // Add it to the list
-            mGeofenceList.add(geofence);
-        }
-    }
-
-    /***
-     * Creates a GeofencingRequest object using the mGeofenceList ArrayList of Geofences
-     * Used by {@code #registerGeofences}
-     *
-     * @return the GeofencingRequest object
-     */
-    private GeofencingRequest getGeofencingRequest() {
-        GeofencingRequest.Builder builder = new GeofencingRequest.Builder();
-        builder.setInitialTrigger(GeofencingRequest.INITIAL_TRIGGER_ENTER);
-        builder.addGeofences(mGeofenceList);
-        return builder.build();
-    }
-
-    private PendingIntent getGeofencePendingIntent() {
-        // Reuse the PendingIntent if we already have it.
-        if (mGeofencePendingIntent != null) {
-            return mGeofencePendingIntent;
-        }
-        Intent intent = new Intent(mContext, GeofenceBroadcastReceiver.class);
-        mGeofencePendingIntent = PendingIntent.getBroadcast(mContext, 0, intent, PendingIntent.
-                FLAG_UPDATE_CURRENT);
-        return mGeofencePendingIntent;
-    }
-}
